refactor: have dedicated `utils` module
