+++ folding-verifier-solidity/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod encoding;
